Refactor: Restructure project
- Move validation method to AppUtil.cpp
- Remove code testing GA.
- Make Dep lib be linked with PTSD instead of GA.
- Update the url to PTSD.
