+++ CMakeLists.txt
@@ -27,8 +27,11 @@ target_include_directories(${PROJECT_NAME} SYSTEM PRIVATE ${DEPENDENCY_INCLUDE_D
 target_include_directories(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/PTSD/include)
 target_include_directories(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
 
-target_link_libraries(${PROJECT_NAME}
+target_link_libraries(Framework
     ${DEPENDENCY_LINK_LIBRARIES}
+)
+
+target_link_libraries(${PROJECT_NAME}
     SDL2::SDL2main
     Framework
 )
